requirements.txt
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
decorator==5.1.1
grpcio==1.65.4
grpcio-tools==1.65.4
idna==3.7
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.0.8
numba==0.60.0
numpy==2.0.1
packaging==24.1
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
PyAudio==0.2.14
pycparser==2.22
requests==2.31.0
scikit-learn==1.5.1
scipy==1.14.1
soundfile==0.12.1
soxr==0.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
webrtcvad==2.0.10
